Downgrade to broadway from gtk3.24.36 (I think, latest 3.x line)
This is to try to maintain compatiblity with applications for
both GTK3 and GTK4, not sure if it will work or not.

GTK3 and GTK4 versions of broadwayd can be built, but I don't
think the output will differ. Both binaries work the same and
can be run in parallel

Note: A future attempt may be made at a winelib driver