Downgrade to broadway from gtk3.24.36 (I think, latest 3.x line)
This is to try to maintain compatiblity with applications for
both GTK3 and GTK4, not sure if it will work or not.

GTK3 and GTK4 versions of broadwayd can be built, but I don't
think the output will differ. Both binaries work the same and
can be run in parallel

Note: A future attempt may be made at a winelib driver

Start working on trying to figure out how to convert gdk engine to wine

Started working on building the broadway client library as standalone
with wine_broadway prefixes on functions to eventually be merged in
to a winebroadway driver or somewhere else.

Compile broadway-server with a minimal amount of Gtk/Gdk private
windowing structures and such.

This will be replaced by Win32 counterparts during port

Disable hashtable lookups for the moment

work on isolating pure broadway code from GDK

Validated standalone build

Update notes, remove gdk dependancy on broadwayd

Checkpoint
 - Change g_memdup to g_memdup2 in broadway server
 - Start working on Makefile rules to build static library
 - Import gdk from gtk+-3.24.36
 - Start working on .gitignore